Fix: So if you use the videos filter and switch to another playlist without clearing the search, then when you use the filter on the that playist's videos and clear the search, the videos in current playlist are basically replaced with the  videos from the previous one
Cause: filteredVideos array is not reset when I switch to another playlist mid-search, the search has to be cleared for unfilteredVideos to become an empty array. So I have to clear the array every time the $route is called.

One line fix but not immediately apparent. I wasn't even going to try to fix this bug but was like what the hell. Glad I didn't postpone it.